metpetdb_api/requirements/dev.txt
Django>=1.11.0,<=1.11.10
django-concurrency>=0.9
django-debug-toolbar>=1.3.2,<=1.4
django-dotenv>=1.3.0
django-extensions>=1.5.5,<=2.0
django-getenv>=1.3.1
djangorestframework-csv>=2.1
djangorestframework>=3.7,<=3.8
psycopg2-binary>=2.6.1
six>=1.9.0
sqlparse>=0.1.15
wheel>=0.24.0
pathlib
django-versatileimagefield
xlrd
